Add basic configuration for git and bash
This is the initial attempt of organizing dotfiles in a repository. Git
configuration was made such as adding messages and setting up vscode as
editor. In bash aliases only update and some ls aliases are added.
An installation script is a also included. This installation script is
safe to run as it backs up existing dotfiles and then creates symlinks
to the dotfiles in this repository.